backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
numpy==1.24.3
scipy==1.11.4
pandas==2.1.3
geopy==2.4.0
python-dotenv==1.0.0
cors==1.0.1
asyncpg==0.29.0
